Split `PublicationCallGateway` by two separate gateways for both post and comments
