--- packages/react/src/transactions/adapters/PublicationCallGateway.ts
@@ -1,345 +1,0 @@
-import { ApolloClient, NormalizedCacheObject } from '@apollo/client';
-import {
-  CollectModuleParams,
-  CreatePublicPostRequest as CreatePublicPostRequestArg,
-  CreatePublicCommentRequest as CreatePublicCommentRequestArg,
-  CreatePostTypedDataDocument,
-  CreatePostTypedDataMutation,
-  CreatePostTypedDataMutationVariables,
-  CreatePostViaDispatcherDocument,
-  CreatePostViaDispatcherMutation,
-  CreatePostViaDispatcherMutationVariables,
-  omitTypename,
-  PublicationMainFocus,
-  PublicationMetadataDisplayTypes,
-  PublicationMetadataMediaInput,
-  PublicationMetadataV2Input,
-  ReferenceModuleParams,
-  CreateCommentViaDispatcherDocument,
-  CreateCommentViaDispatcherMutation,
-  CreateCommentViaDispatcherMutationVariables,
-  CreateCommentTypedDataMutationVariables,
-  CreateCommentTypedDataDocument,
-  CreateCommentTypedDataMutation,
-} from '@lens-protocol/api-bindings';
-import {
-  NativeTransaction,
-  Nonce,
-  TransactionError,
-  TransactionErrorReason,
-  TransactionKind,
-} from '@lens-protocol/domain/entities';
-import {
-  ChargeCollectPolicy,
-  CollectPolicyType,
-  CreatePostRequest,
-  CreateCommentRequest,
-  ICreatePostCallGateway,
-  Media,
-  NftAttribute,
-  ReferencePolicy,
-} from '@lens-protocol/domain/use-cases/publications';
-import { SupportedTransactionRequest } from '@lens-protocol/domain/use-cases/transactions';
-import { ChainType, failure, invariant, success } from '@lens-protocol/shared-kernel';
-import { v4 } from 'uuid';
-
-import { UnsignedLensProtocolCall } from '../../wallet/adapters/ConcreteWallet';
-import { AsyncRelayReceipt, ITransactionFactory } from './ITransactionFactory';
-import { UploadHandler } from './UploadHandler';
-
-function mapMetaAttributes(attributes: NftAttribute[]) {
-  return attributes.map(({ displayType, value, ...rest }) => {
-    return {
-      ...rest,
-      value: value.toString(),
-      displayType: PublicationMetadataDisplayTypes[displayType],
-    };
-  });
-}
-
-function mapMedia(media: Media): PublicationMetadataMediaInput {
-  return {
-    item: media.url,
-    type: media.mimeType,
-  };
-}
-
-function createPublicationMetadata(
-  request: CreatePostRequest | CreateCommentRequest,
-): PublicationMetadataV2Input {
-  const sharedMetadata = {
-    version: '2.0.0',
-    metadata_id: v4(),
-
-    content: request.content,
-    media: request.media?.map(mapMedia),
-    locale: request.locale,
-    mainContentFocus: PublicationMainFocus[request.contentFocus],
-  };
-
-  switch (request.collect.type) {
-    case CollectPolicyType.CHARGE:
-    case CollectPolicyType.FREE:
-      return {
-        ...sharedMetadata,
-
-        attributes: mapMetaAttributes(request.collect.metadata.attributes),
-        description: request.collect.metadata.description,
-        name: request.collect.metadata.name,
-      };
-
-    case CollectPolicyType.NO_COLLECT:
-      return {
-        ...sharedMetadata,
-
-        name: 'none', // although "name" is not needed when a publication is not collectable, out Publication Metadata V2 schema requires it ¯\_(ツ)_/¯
-        attributes: [],
-      };
-  }
-}
-
-function resolveCollectModuleFeeParams(collect: ChargeCollectPolicy) {
-  return {
-    amount: {
-      currency: collect.fee.asset.address,
-      value: collect.fee.toSignificantDigits(),
-    },
-    followerOnly: collect.followersOnly,
-    recipient: collect.recipient,
-    referralFee: collect.mirrorReward,
-  };
-}
-
-function resolveChargeCollectModule(collect: ChargeCollectPolicy): CollectModuleParams {
-  if (collect.collectLimit && collect.timeLimited) {
-    return {
-      limitedTimedFeeCollectModule: {
-        ...resolveCollectModuleFeeParams(collect),
-        collectLimit: collect.collectLimit.toString(),
-      },
-    };
-  }
-  if (collect.collectLimit) {
-    return {
-      limitedFeeCollectModule: {
-        ...resolveCollectModuleFeeParams(collect),
-        collectLimit: collect.collectLimit.toString(),
-      },
-    };
-  }
-  if (collect.timeLimited) {
-    return {
-      timedFeeCollectModule: resolveCollectModuleFeeParams(collect),
-    };
-  }
-  return {
-    feeCollectModule: resolveCollectModuleFeeParams(collect),
-  };
-}
-
-function resolveCollectModule(
-  request: CreatePostRequest | CreateCommentRequest,
-): CollectModuleParams {
-  switch (request.collect.type) {
-    case CollectPolicyType.FREE:
-      return {
-        freeCollectModule: {
-          followerOnly: request.collect.followersOnly,
-        },
-      };
-
-    case CollectPolicyType.NO_COLLECT:
-      return {
-        revertCollectModule: true,
-      };
-
-    case CollectPolicyType.CHARGE:
-      return resolveChargeCollectModule(request.collect);
-  }
-}
-
-function resolveReferenceModule(
-  request: CreatePostRequest | CreateCommentRequest,
-): ReferenceModuleParams | undefined {
-  if (request.reference === ReferencePolicy.FOLLOWERS_ONLY) {
-    return {
-      followerOnlyReferenceModule: true,
-    };
-  }
-  return undefined;
-}
-
-export class PublicationCallGateway implements ICreatePostCallGateway {
-  constructor(
-    private readonly apolloClient: ApolloClient<NormalizedCacheObject>,
-    private readonly transactionFactory: ITransactionFactory<SupportedTransactionRequest>,
-    private readonly upload: UploadHandler,
-  ) {}
-
-  async createDelegatedTransaction<T extends CreatePostRequest | CreateCommentRequest>(
-    request: T,
-  ): Promise<NativeTransaction<T>> {
-    return this.transactionFactory.createNativeTransaction({
-      chainType: ChainType.POLYGON,
-      id: v4(),
-      request,
-      asyncRelayReceipt: this.initiatePublicationCreation(request),
-    });
-  }
-
-  async createUnsignedProtocolCall<T extends CreateCommentRequest>(
-    request: T,
-    nonce?: Nonce,
-  ): Promise<UnsignedLensProtocolCall<T>>;
-  async createUnsignedProtocolCall<T extends CreatePostRequest>(
-    request: T,
-    nonce?: Nonce,
-  ): Promise<UnsignedLensProtocolCall<T>>;
-  async createUnsignedProtocolCall(
-    request: CreateCommentRequest | CreatePostRequest,
-    nonce?: Nonce,
-  ): Promise<UnsignedLensProtocolCall<CreateCommentRequest | CreatePostRequest>> {
-    switch (request.kind) {
-      case TransactionKind.CREATE_POST:
-        return this.createPostCall(request, nonce);
-
-      case TransactionKind.CREATE_COMMENT:
-        return this.createCommentCall(request, nonce);
-    }
-  }
-
-  private async initiatePublicationCreation(
-    request: CreatePostRequest | CreateCommentRequest,
-  ): AsyncRelayReceipt {
-    switch (request.kind) {
-      case TransactionKind.CREATE_POST:
-        return this.initiatePostCreation(request);
-
-      case TransactionKind.CREATE_COMMENT:
-        return this.initiateCommentCreation(request);
-    }
-  }
-
-  private async initiatePostCreation(request: CreatePostRequest): AsyncRelayReceipt {
-    const { data } = await this.apolloClient.mutate<
-      CreatePostViaDispatcherMutation,
-      CreatePostViaDispatcherMutationVariables
-    >({
-      mutation: CreatePostViaDispatcherDocument,
-      variables: {
-        request: await this.resolveCreatePostRequestArg(request),
-      },
-    });
-
-    invariant(data, 'Cannot create post via dispatcher');
-
-    if (data.result.__typename === 'RelayError') {
-      return failure(new TransactionError(TransactionErrorReason.REJECTED));
-    }
-
-    return success({
-      indexingId: data.result.txId,
-      txHash: data.result.txHash,
-    });
-  }
-
-  private async initiateCommentCreation(request: CreateCommentRequest): AsyncRelayReceipt {
-    const { data } = await this.apolloClient.mutate<
-      CreateCommentViaDispatcherMutation,
-      CreateCommentViaDispatcherMutationVariables
-    >({
-      mutation: CreateCommentViaDispatcherDocument,
-      variables: {
-        request: await this.resolveCreateCommentRequestArg(request),
-      },
-    });
-
-    invariant(data, 'Cannot create comment via dispatcher');
-
-    if (data.result.__typename === 'RelayError') {
-      return failure(new TransactionError(TransactionErrorReason.REJECTED));
-    }
-
-    return success({
-      indexingId: data.result.txId,
-      txHash: data.result.txHash,
-    });
-  }
-
-  private async createPostCall<T extends CreatePostRequest>(
-    request: T,
-    nonce?: Nonce,
-  ): Promise<UnsignedLensProtocolCall<T>> {
-    const { data } = await this.apolloClient.mutate<
-      CreatePostTypedDataMutation,
-      CreatePostTypedDataMutationVariables
-    >({
-      mutation: CreatePostTypedDataDocument,
-      variables: {
-        request: await this.resolveCreatePostRequestArg(request),
-        options: nonce ? { overrideSigNonce: nonce } : undefined,
-      },
-    });
-
-    invariant(data, 'Cannot generate typed data for post creation');
-
-    return new UnsignedLensProtocolCall(
-      data.result.id,
-      request,
-      omitTypename(data.result.typedData),
-    );
-  }
-
-  private async createCommentCall<T extends CreateCommentRequest>(
-    request: T,
-    nonce?: Nonce,
-  ): Promise<UnsignedLensProtocolCall<T>> {
-    const { data } = await this.apolloClient.mutate<
-      CreateCommentTypedDataMutation,
-      CreateCommentTypedDataMutationVariables
-    >({
-      mutation: CreateCommentTypedDataDocument,
-      variables: {
-        request: await this.resolveCreateCommentRequestArg(request),
-        options: nonce ? { overrideSigNonce: nonce } : undefined,
-      },
-    });
-
-    invariant(data, 'Cannot generate typed data for comment creation');
-
-    return new UnsignedLensProtocolCall(
-      data.result.id,
-      request,
-      omitTypename(data.result.typedData),
-    );
-  }
-
-  private async resolveCreatePostRequestArg(
-    request: CreatePostRequest,
-  ): Promise<CreatePublicPostRequestArg> {
-    return {
-      profileId: request.profileId,
-      contentURI: await this.uploadPublicationMetadata(request),
-      collectModule: resolveCollectModule(request),
-      referenceModule: resolveReferenceModule(request),
-    };
-  }
-
-  private async resolveCreateCommentRequestArg(
-    request: CreateCommentRequest,
-  ): Promise<CreatePublicCommentRequestArg> {
-    return {
-      profileId: request.profileId,
-      publicationId: request.publicationId,
-      contentURI: await this.uploadPublicationMetadata(request),
-      collectModule: resolveCollectModule(request),
-      referenceModule: resolveReferenceModule(request),
-    };
-  }
-
-  private async uploadPublicationMetadata(
-    request: CreatePostRequest | CreateCommentRequest,
-  ): Promise<string> {
-    return this.upload(createPublicationMetadata(request));
-  }
-}
